Text2TextGenerationPipeline, TranslationPipeline, SummarizationPipeline
